+++ securitytools/SecurityToolsCore/Store/ICertificateStore.cs
@@ -9,6 +9,7 @@ namespace GGolbik.SecurityTools.Store;
 public interface ICertificateStore
 {
     bool MoveCertificateTo(string thumbprint, ICertificateStore store);
+    void UpdatePassword(byte[]? password, PbeParameters? pbeParameters = null);
 
     #region Get
     IList<X509Certificate2> GetCertificates();
@@ -17,8 +18,15 @@ public interface ICertificateStore
     X509Certificate2? GetCertificate(string thumbprint, bool includePrivateKey);
     IList<AsymmetricAlgorithm> GetKeyPairs();
     AsymmetricAlgorithm? GetKeyPair(string thumbprint);
+    /// <summary>
+    /// Returns a Thumbprints of key pairs which could not be read.
+    /// A reason might be an invalid password.
+    /// </summary>
+    /// <returns>The thumbprints.</returns>
+    IList<string> GetKeyPairsWithError();
     IList<X509Crl> GetCrls();
     IList<X509Crl> GetCrlsOfIssuer(string thumbprint);
+    IList<X509Crl> GetCrlsOfIssuer(string thumbprint, bool verify);
     X509Crl? GetCrl(string thumbprint);
     IList<CertificateRequest> GetCsrs();
     CertificateRequest? GetCsr(string thumbprint);