+++ securitytools/SecurityToolsCore/Work/WorkState.cs
@@ -1,0 +1,11 @@
+
+namespace GGolbik.SecurityTools.Work;
+
+public enum WorkState
+{
+    None = 0,
+    Waiting = 1,
+    Executing = 2,
+    Error = 3,
+    Done = 4
+}