+++ securitytools/SecurityToolsTest/Helper/ICertificateStoreFactory.cs
@@ -1,0 +1,10 @@
+using System.Security.Cryptography;
+using GGolbik.SecurityTools.Store;
+
+namespace GGolbik.SecurityToolsTest;
+
+public interface ICertificateStoreFactory : IDisposable
+{
+    ICertificateStore Create();
+    ICertificateStore Create(byte[] password, PbeParameters? pbeParameters = null);
+}