+++ securitytools/SecurityToolsTest/Tests/InMemCertificateStoreTest.cs
@@ -5,24 +5,10 @@ namespace GGolbik.SecurityToolsTest;
 public class InMemCertificateStoreTest : CertificateStoreTest
 {
 
-    private class InMemCertificateStoreFactory : ICertificateStoreFactory
-    {
-        public InMemCertificateStoreFactory()
-        {
-        }
-        public ICertificateStore Create()
-        {
-            return new InMemCertificateStore();
-        }
-
-        public void Dispose()
-        {
-        }
-    }
-
     public InMemCertificateStoreTest() : base(new InMemCertificateStoreFactory())
     {
     }
+
     [Fact]
     public override void Test_Add_CertificateDer()
     {